Structure file and specs
